Moved entities. Added necessary dependencies to project.
